Update docs(types.ts) fix pos and area in read()
After removing scale, one of the sprites pos needed to be adjusted to overlap other sprite in demonstration. 
Also, area() is needed as well, because without area(), kaboom will throw an error "purpleBean is not a function" when using "purpleBean.onClick()".
Not sure if this is a bug or intended like that, but might explain one of the reasons why people get errors often.